Combine players_list, num_players_online into get_info_from_server
New serverInfo namedtuple was created to store both the players list and number of online players.
players will be an empty list if no players are online.

Fixed previous issue of players_list function which threw an error because server.status().raw
did not have a list of players if there was no one online.

Updated functions in looped_tasks.py to reflect these changes.