Delete !online command and bot response after cooldown
When someone uses the !online command, their command and the response from the bot will now be deleted after a cooldown (1 as of right now).

This "cooldown" is simply one less than the number of rotations the bot does of its activity status.

This was done because the !online messages would become outdated by nature (as players join and leave) and inaccurate responses were cluttering the channels.

 update_activity method was also changed to support this functionality.